requirements.txt
azure-functions
azure-ai-projects>=2.0.0b1
azure-identity
requests
pyyaml
streamlit
